Update Perfect Number
Problem Statement- 
Given a number N. Check if it is perfect or not. A number is perfect if sum of factorial of its digit is equal to the given number.

Example 1:

Input: N = 23
Output: 0
Explanation: The sum of factorials of 
digits of 23 is 2! + 3! = 2 + 6 = 8 
which is not equal to 23. Thus, answer 
is 0.